+++ packages/node-core/src/indexer/project.service.ts
@@ -20,7 +20,7 @@ import {PoiSyncService} from './poi';
 import {PoiService} from './poi/poi.service';
 import {StoreService} from './store.service';
 import {isCachePolicy} from './storeCache';
-import {ISubqueryProject, IProjectService} from './types';
+import {ISubqueryProject, IProjectService, BypassBlocks} from './types';
 import {IUnfinalizedBlocksService} from './unfinalizedBlocks.service';
 
 const logger = getLogger('Project');
@@ -84,11 +84,15 @@ export abstract class BaseProjectService<
     return this._blockOffset;
   }
 
+  get bypassBlocks(): BypassBlocks {
+    return this.project.network.bypassBlocks ?? [];
+  }
+
   protected get isHistorical(): boolean {
     return this.storeService.historical;
   }
 
-  private async getExistingProjectSchema(): Promise<string | undefined> {
+  protected async getExistingProjectSchema(): Promise<string | undefined> {
     return getExistingProjectSchema(this.nodeConfig, this.sequelize);
   }
 
@@ -133,14 +137,13 @@ export abstract class BaseProjectService<
         void this.poiSyncService.syncPoi(undefined);
       }
 
+      const reindexedUpgrade = await this.initUpgradeService(this.startHeight);
       // Unfinalized is dependent on POI in some cases, it needs to be init after POI is init
       const reindexedUnfinalized = await this.initUnfinalizedInternal();
 
       if (reindexedUnfinalized !== undefined) {
         this._startHeight = reindexedUnfinalized;
       }
-
-      const reindexedUpgrade = await this.initUpgradeService(this.startHeight);
 
       if (reindexedUpgrade !== undefined) {
         this._startHeight = reindexedUpgrade;