+++ packages/node-core/CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [14.1.7] - 2024-10-30
+### Changed
+- Bump `@subql/common` dependency
+
+## [14.1.6] - 2024-10-21
+### Fixed
+- Issues with setting a large block range for bypass blocks (#2566)
+- Test runner not setting lastProcessedHeight leading to data not being flushed (#2569)
+- Unable to rewind unfinalized blocks on startup (#2570)
+- Store `getByFields` returning removed cache data (#2571)
+
+### Changed
+- Throw error when store getByField(s) options.limit exceeds queryLimit option (#2567)
+
 ## [14.1.5] - 2024-09-25
 ### Changed
 - Bump common, Added manifest support for query-subgraph.
@@ -826,7 +840,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move blockchain agnostic code from `node` to `node-core` package. (#1222)
 
-[Unreleased]: https://github.com/subquery/subql/compare/node-core/14.1.5...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node-core/14.1.7...HEAD
+[14.1.7]: https://github.com/subquery/subql/compare/node-core/14.1.6...node-core/14.1.7
+[14.1.6]: https://github.com/subquery/subql/compare/node-core/14.1.5...node-core/14.1.6
 [14.1.5]: https://github.com/subquery/subql/compare/node-core/14.1.4...node-core/14.1.5
 [14.1.4]: https://github.com/subquery/subql/compare/node-core/14.1.3...node-core/14.1.4
 [14.1.3]: https://github.com/subquery/subql/compare/node-core/14.1.2...node-core/14.1.3