+++ packages/node/CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.9] - 2024-10-30
+### Changed
+- Bump `@subql/node-core` dependency
+
+## [5.2.8] - 2024-10-23
+### Changed
+- Bump `@subql/common-substrate` dependency
+
+## [5.2.7] - 2024-10-21
+### Fixed
+- Issues with setting a large block range for bypass blocks (#2566)
+
 ## [5.2.6] - 2024-09-25
 ### Changed
 - Bump common, Added manifest support for query-subgraph.
@@ -1311,7 +1323,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/5.2.6...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node/5.2.9...HEAD
+[5.2.9]: https://github.com/subquery/subql/compare/node/5.2.8...node/5.2.9
+[5.2.8]: https://github.com/subquery/subql/compare/node/5.2.7...node/5.2.8
+[5.2.7]: https://github.com/subquery/subql/compare/node/5.2.6...node/5.2.7
 [5.2.6]: https://github.com/subquery/subql/compare/node/5.2.5...node/5.2.6
 [5.2.5]: https://github.com/subquery/subql/compare/node/5.2.3...node/5.2.5
 [5.2.3]: https://github.com/subquery/subql/compare/node/5.2.2...node/5.2.3