+++ packages/types-core/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-10-21
+### Changed
+- Update store interface making options.limit required on getByField(s) methods (#2567)
+
 ## [1.1.1] - 2024-08-12
 ### Fixed
 - DS Processor types not being able to distinguish input and filter types (#2522)
@@ -70,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of new package containing types common to all chains
 
-[Unreleased]: https://github.com/subquery/subql/compare/types-core/1.1.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/types-core/2.0.0...HEAD
+[2.0.0]: https://github.com/subquery/subql/compare/types-core/1.1.1...types-core/2.0.0
 [1.1.1]: https://github.com/subquery/subql/compare/types-core/1.1.0...types-core/1.1.1
 [1.1.0]: https://github.com/subquery/subql/compare/types-core/1.0.0...types-core/1.1.0
 [1.0.0]: https://github.com/subquery/subql/compare/types-core/0.10.0...types-core/1.0.0