+++ packages/types/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.11.3] - 2024-10-23
+### Fixed
+- Bump version `@subql/types-core`
+
 ## [3.11.2] - 2024-08-14
 ### Added
 - Update polkadot/api library
@@ -267,7 +271,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support block handler
 
-[Unreleased]: https://github.com/subquery/subql/compare/types/3.11.2...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/types/3.11.3...HEAD
+[3.11.3]: https://github.com/subquery/subql/compare/types/3.11.2...types/3.11.3
 [3.11.2]: https://github.com/subquery/subql/compare/types/3.11.1...types/3.11.2
 [3.11.1]: https://github.com/subquery/subql/compare/types/3.11.0...types/3.11.1
 [3.11.0]: https://github.com/subquery/subql/compare/types/3.10.0...types/3.11.0